Extract SortDirection to a top-level class
